refactor(App, Calculator, Console): 기능별 클래스 별도 파일로 분리
